gradle fixes and basic city item appearance

+++ common/feature/city/data/src/main/java/com/city/data/local/repository/CityLocalRoomRepositoryImpl.kt
@@ -35,7 +35,7 @@ class CityLocalRoomRepositoryImpl(
     override suspend fun modifyCity(city: City): CitiesModifyResult {
         return try {
             CitiesModifyResult.Success(
-                database.cityDao.updateCity(city.toCityEntity())
+                database.cityDao.updateCity(city.toCityEntity()).toLong()
             )
         } catch (e: Exception) {
             CitiesModifyResult.Failure
